Default Inputs type, move validation comparison functions
