+++ src/hook.ts
@@ -2,13 +2,7 @@ import { useReducer, useRef } from 'react';
 import debounce from 'lodash.debounce';
 import { nanoid } from 'nanoid/non-secure';
 
-import {
-  ValidationProps,
-  Oxin,
-  ValidationState,
-  Validator,
-  ValidatorTuple,
-} from './types';
+import { ValidationProps, Oxin, ValidationState } from './types';
 
 import { createInitialState, OxinReducer, createReducer } from './reducer';
 import { setValue, removeField, setValidation } from './actions';
@@ -16,6 +10,8 @@ import {
   runValidators,
   getBooleanValidators,
   mergeValidators,
+  validationEquals,
+  validatorsEquals,
 } from './validation';
 
 interface Cache {
@@ -36,24 +32,7 @@ function useCache(): Cache {
   return { getOrSet, set, has, get };
 }
 
-// TODO: move to validation
-const validationEquals = (v1?: ValidationState, v2?: ValidationState) => {
-  const stringify = (obj: ValidationState) =>
-    Object.values(obj)
-      .map((val) => JSON.stringify(val))
-      .join('');
-
-  return !v1 || !v2 ? false : stringify(v1) === stringify(v2);
-};
-
-const validatorsEquals = (
-  v1: (Validator | ValidatorTuple)[],
-  v2: (Validator | ValidatorTuple)[],
-) => {
-  return JSON.stringify(v1) === JSON.stringify(v2);
-};
-
-export function useOxin<Inputs>(): Oxin<Inputs> {
+export function useOxin<Inputs = Record<string, unknown>>(): Oxin<Inputs> {
   const [inputState, dispatch] = useReducer<OxinReducer<Inputs>>(
     createReducer<Inputs>(),
     createInitialState<Inputs>(),
